+++ resources/database.rb
@@ -1,32 +1,15 @@
-#
-# Cookbook:: postgresql
-# Resource:: database
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
 unified_mode true
+use 'partial/_connection'
+
+# Override the database property from the partial
+property :database,
+         String,
+         name_property: true
 
 property :template, String, default: 'template1'
 property :encoding, String
 property :locale,   String
 property :owner,    String
-
-# Connection prefernces
-property :user,     String, default: 'postgres'
-property :database, String, name_property: true
-property :host,     String
-property :port,     Integer, default: 5432
-property :psqlrc,   [true, false], default: true
 
 action :create do
   createdb = 'createdb'